LayoutTests/platform/ios/ios/fast/coordinates/element-client-rects.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
<script src="resources/helpers.js"></script>
<style>* { margin: 0; padding: 0; }</style>
</head>
<body>
<div id="box" style="background:blue;width:100px;height:100px;position:absolute;top:100px;left:100px"></div>
<p id="description"></p>
<div id="console"></div>
<!-- Filler to allow scrolling -->
<div style="width:1500px;height:1500px;"></div>
<script>
description("This tests Element.getBoundingClientRects and getClientRects positions when unscaled, scaled, and panned.");

const scale = 2;
const panX = 10;
const panY = 10;

var box = document.getElementById('box');

debug('');
debug('unscaled');
shouldBeEqualToString("rectString(document.getElementById('box').getBoundingClientRect())", "(100,100 100x100)");
shouldBeEqualToString("rectStrings(document.getElementById('box').getClientRects())", "(100,100 100x100)");

debug('');
debug('scaled and panned');
setInitialScaleAndPanBy(scale, panX, panY);
var rect = "(" + (100-panX) + "," + (100-panY) + " 100x100)";
shouldBeEqualToString("rectString(document.getElementById('box').getBoundingClientRect())", rect);
shouldBeEqualToString("rectStrings(document.getElementById('box').getClientRects())", rect);

debug('');

var successfullyParsed = true;
endTest();
</script>
</body>
</html>
