LayoutTests/platform/ios/ios/fast/coordinates/element-client-rects-expected.txt

This tests Element.getBoundingClientRects and getClientRects positions when unscaled, scaled, and panned.
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

unscaled
PASS rectString(document.getElementById('box').getBoundingClientRect()) is "(100,100 100x100)"
PASS rectStrings(document.getElementById('box').getClientRects()) is "(100,100 100x100)"

scaled and panned
PASS rectString(document.getElementById('box').getBoundingClientRect()) is "(90,90 100x100)"
PASS rectStrings(document.getElementById('box').getClientRects()) is "(90,90 100x100)"

PASS successfullyParsed is true

TEST COMPLETE
